requirements.txt
aiosmtpd==1.3.0
amqp==5.1.1
asgiref==3.5.2
astroid==2.13.2
async-timeout==4.0.2
atpublic==2.3
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
black==21.10b0
boto3==1.19.2
botocore==1.22.2
build==0.9.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
cssbeautifier==1.14.7
ddt==1.6.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dj-database-url==0.5.0
django-allauth==0.52.0
Django==4.1.9
django-customselectwidget==0.1.1
django-dotenv==1.4.2
django-extensions==3.1.3
django-heroku==0.3.1
django-json-widget==1.1.1
django-object-actions==4.1.0
django-organizations==2.0.1
django-phonenumber-field==5.2.0
django-sendgrid-v5==1.2.1
django-simple-history==3.0.0
django-sites==0.11
django-storages==1.12.2
django-tables2==2.4.1
djlint==1.19.13
EditorConfig==0.12.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
future==0.18.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
jmespath==0.10.0
jsbeautifier==1.14.7
kombu==5.2.4
lazy-object-proxy==1.7.1
Levenshtein==0.20.8
mailchimp-transactional==1.0.47
mccabe==0.7.0
mypy==0.910
mypy-extensions==0.4.3
nullsmtpd==0.5.1
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pathspec==0.10.3
pep517==0.13.0
phonenumbers==8.12.35
pip-tools==6.12.1
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.7
python-Levenshtein==0.20.8
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
rapidfuzz==2.13.2
redis==4.4.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.5.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.1.0
thefuzz==0.19.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.11.6
tqdm==4.64.1
types-pytz==2022.7.0.0
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
tzdata==2022.7
Unipath==1.1
urllib3==1.26.7
uuid7==0.1.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
wrapt==1.14.0
XlsxWriter==3.0.3
